fix(writer): cast step to i64 in bpb_sample ActiveModel (Wave 24)
- neon_writer::bpb_sample: step: Set(step) -> Set(step as i64)
  to match the new BIGINT column type
- tests/ledger_seaorm.rs: step: i32 = 200 -> step: i64 = 200
  to match updated Model type

Public function signature bpb_sample(seed: i32, step: i32, ...) is
unchanged for callers; the cast happens internally.

Anchor: phi^2 + phi^-2 = 3 · DOI 10.5281/zenodo.19227877
Refs #118

+++ tests/ledger_seaorm.rs
@@ -50,7 +50,7 @@ async fn ledger_seaorm_smoke() {
 
     let canon = "ledger_seaorm_smoke_test";
     let seed: i64 = 47; // BIGINT
-    let step: i32 = 200;
+    let step: i64 = 200; // BIGINT (Wave 24: step lifted from INT to BIGINT)
     let bpb: f64 = 2.19;
 
     // Insert (idempotent: DO NOTHING on conflict).
